pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mutual-fund-backend"
version = "1.0.0"
description = "API backend for managing mutual fund data and operations"
authors = [{ name = "Hemanth Kumar Pasham, email = 'hemanthkumarpasham9502@gmail.com" }]
readme = "README.md"
requires-python = ">= 3.12"
# Unpinned Dependencies below...
# [Run pip-compile (or uv pip compile) to get the pinned versions in requirements.txt]
dependencies = [
    "fastapi[all]>=0.115.14",
    "uvicorn>=0.22.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.9.0",
    "sqlalchemy>=2.0.29",
    "sqlalchemy-utils>=0.37.8",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.8.0",
    "passlib>=1.7.4",
    "bcrypt>=3.2.0"
]

[tool.ruff]
line-length = 88  # Customize line length
target-version = "py312"  # Adjust to your Python version

[tool.ruff.format]
quote-style = "double"  # Use double quotes
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = ["E", "F", "B", "W"]  # Example: Enable error (E), flake8 (F), bugbear (B), warnings (W)
ignore = ["E501", "S608"]  # Ignore line length warnings (Black handles this)

[tool.black]
line-length = 88
target-version = ['py311']
skip-string-normalization = false  # Keep string quotes consistent


[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["bot_common", "bot_common.*"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=20"
testpaths = [
    "test"
]
python_files = [
    "test_*.py"
]
python_functions = [
    "test_*"
]
